+++ crates/utils/src/tracing/grpc.rs
@@ -46,6 +46,7 @@ fn rpc_trace_fn<T>(request: &http::Request<T>) -> tracing::Span {
         Some("CheckNullifiers") => rpc_span!("rpc.rpc", "CheckNullifiers"),
         Some("CheckNullifiersByPrefix") => rpc_span!("rpc.rpc", "CheckNullifiersByPrefix"),
         Some("GetBlockHeaderByNumber") => rpc_span!("rpc.rpc", "GetBlockHeaderByNumber"),
+        Some("SyncStorageMaps") => rpc_span!("rpc.rpc", "SyncStorageMaps"),
         Some("SyncState") => rpc_span!("rpc.rpc", "SyncState"),
         Some("SyncNotes") => rpc_span!("rpc.rpc", "SyncNotes"),
         Some("GetNotesById") => rpc_span!("rpc.rpc", "GetNotesById"),